HW_4

+++ VeterenaryClinic/Models/ViewModels/PriceViewModel.cs
@@ -9,7 +9,7 @@ namespace VeterenaryClinic.Models.ViewModels
     public class PriceViewModel
     {
         public int Id { get; set; }
-        public decimal PriceValue { get; set; }
+        public double PriceValue { get; set; }
         public ICollection<VeterenaryClinicVievModel> VetClinics { get; set; }
     }
 }
